requirements.txt
PySide6==6.10.1
Pillow==12.1.0
pynput==1.7.6
keyboard==0.13.5
pyinstaller==6.8.0